Rename databases-ui.ts to local-databases-ui.ts
