+++ extensions/ql-vscode/src/extension.ts
@@ -44,7 +44,7 @@ import {
 } from "./config";
 import { install } from "./languageSupport";
 import { DatabaseItem, DatabaseManager } from "./local-databases";
-import { DatabaseUI } from "./databases-ui";
+import { DatabaseUI } from "./local-databases-ui";
 import {
   TemplatePrintAstProvider,
   TemplatePrintCfgProvider,
