+++ extensions/ql-vscode/src/local-databases-ui.ts
